Sanity-check length before extracting sub-regions
Test whether the length of the subregion to be extracted from the
total protein sequence has a length of one or above. This avoids
warnings which can arise when a DNA sequence does not end cleanly on
the codon boundary of the last codon of a region.

Turn off logging to database.

Adapt formating and identation to Perl standard.