Accelerated and TCR-enabled read mapping
Mapping of dual-tagged reads to physical positions is often the
most time consuming step of data processing and does not
parallelize well due to locks on database tables. In addition,
in the current version an individual instance of the script does
not restrict itself to the locus that was initially passed to it
via the command line. This can create non-deterministic behavior,
when processing of multiple loci is parallelized. Finally also this
script requires modification for TCR processing. Changes:

Collect all seq_ids for a given well_id or consensus_id, then
perform the SQL UPDATE in a single statement (currently each
line is updated individually).

Fix the SQL statement selecting unprocessed well_id and include
the assigned locus as criterium.

Add new locus suffixes for the well_id string to facilitate
processing of TCR sequences.