Fix bug in metainfo import
Currently the todb_sampleinfo_highth.pl script assumes that each
identfier used in *_metainfo.tsv is also present in *_plate.tsv
and crashes if this assumption is not met. Add condition to
catch such an event and present a warning to the user.

Fix minor mistakes in debugging output and comments.