Fix QC handling of invalid locus calls
Currently the tag aggregation routine of the QC module assumes that
only valid (i.e. expected) loci are found in the database and both
the proximal and the distal set of tags contain identical loci.
However, this might not necessarily be the case due to low-
confidence BLAST hits or cross-contaminations.
Fix handling of locus information in QC. Fix log levels in QC.