Fix database INSERTs with long data
Currently, the insert procedures into the 'reads' and the
'log_table' table rely on the silent truncation of their 'name'
and 'command' columns by the database, respectively. This creates
two problems:

1) Starting in version 5.7.5 MySQL has set the mode
"STRICT_TRANS_TABLES" as default. Therefore INSERTs with data longer
then the field length will not be truncated anymore but fail and
produce an error.

2) Read IDs typically contain a common prefix that is shared by all
reads of a sequencing run. For long read IDs the truncation after
the first 45 chars can remove the actually distinct/unique part of
the read ID, thereby creating collisions as the read ID must be
unique. In this case not all reads will be inserted into the
database.

Truncate string for `log_table`.`command` to the first 100 chars.
Truncate string for `reads`.`name` to the last 45 chars.