Add sanity check for filenames and metadata
Currently there are no checks that the runname provided on the
commandline matches the one in the ".info" metadata files. This
produces crashes during the QC that are hard to understand/debug
if this problem is unknown upfront. Extend sanity checks in
"pre-flight.sh" to cover this and include them into the normal
run process by default.